+++ src/lightwin/beam_calculation/tracewin/specs.py
@@ -19,6 +19,7 @@ from lightwin.config.helper import find_file
 from lightwin.config.key_val_conf_spec import KeyValConfSpec
 from lightwin.config.table_spec import TableConfSpec
 from lightwin.constants import example_ini, example_machine_config
+from lightwin.core.elements.field_maps.cavity_settings import EXPORT_PHASES
 
 _PURE_TRACEWIN_CONFIG = (
     KeyValConfSpec(
@@ -269,6 +270,20 @@ _PURE_TRACEWIN_CONFIG = (
         types=(float,),
         description="Input YY’ emittance (mm.mrad) of second beam",
         default_value=-1.0,
+        is_mandatory=False,
+    ),
+    KeyValConfSpec(
+        key="export_phase",
+        types=(str,),
+        description=(
+            "The type of phases that should be exported in the final DAT "
+            "file. Note that ``'as_in_original_dat'`` is not implemented "
+            "yet, but ``'as_in_settings'`` should behave the same way, "
+            "provided that you alter no FieldMap.CavitySettings.reference "
+            "attribute."
+        ),
+        default_value="as_in_settings",
+        allowed_values=EXPORT_PHASES,
         is_mandatory=False,
     ),
     KeyValConfSpec(